mptcp: mib: count MPJ with backup flag
Without such counters, it is difficult to easily debug issues with MPJ
not having the backup flags on production servers.

This is not strictly a fix, but it eases to validate the following patch
without requiring to take packet traces, to query ongoing connections
with Netlink with admin permissions, or to guess by looking at the
behaviour of the scheduler.

Fixes: 4596a2c1b7f5 ("mptcp: allow creating non-backup subflows")
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>
Message-Id: <20240716-mptcp-backup-mpj-v2-4-4d50247405fb@kernel.org>