selftests: mptcp: join: cannot rm sf if closed
Thanks to the previous commit, the MPTCP subflows are now closed on both
directions even when only the MPTCP path-manager of one peer asks for
their closure.

In the two tests modified here -- "userspace pm add & remove address"
and "userspace pm create destroy subflow" -- one peer is controlled by
the userspace PM, and the other one by in-kernel PM. When the userspace
PM sends a RM_ADDR notification, the in-kernel PM will automatically
react by closing all subflows using this address. Now that the subflows
are properly closed on both direction, the userspace PM can no longer
closes the same subflows if they are already closed. In this case, an
error will be returned to the userspace. So no need to run this command,
which mean that the linked counters will then not be incremented.

These tests are then no longer sending both a RM_ADDR, then closing the
linked subflow just after. The test with the userspace PM on the server
side is now removing one subflow linked to one address, then sending
a RM_ADDR for another address. The test with the userspace PM on the
client side is now only removing the subflow that was previously
created.

Fixes: 4369c198e599 ("selftests: mptcp: test userspace pm out of transfer")
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>
Message-Id: <20240802-mptcp-pm-avail-v6-10-964ba9ce279f@kernel.org>