Squash to "mptcp: pm: send ACK on non stale subflows"
If there are only "staled" subflows, pick the first one we found (Mat).

Note that we could also check:

  !subflow>backup && !subflow->request_bkup

It's only a ACK, maybe OK to send that on the backup paths?

If we want this, should we have two alternatives?

  - First try on backup if any ; if not, on "staled" ones?

  - Or just on one of them?

We can also fill an issue, and wait for feedback.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>
Message-Id: <20240807143438.1471057-2-matttbe@kernel.org>