Squash to "mptcp: pm: reuse ID 0 after delete and re-add"
Set the address ID to 0 before calling fill_remote_addresses_vec(): for
fullmesh cases, a bitmap will be created after having looked at all
subflow IDs matching the local one. The ID visible on the wire (e.g. 0)
should be compared to, not the one of the global endpoint (cannot be 0).

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>
Message-Id: <20240815-mptcp-dup-close-evt-v1-3-5a551d3a66cc@kernel.org>