Squash to "selftests: mptcp: join: check re-re-adding ID 0 signal"
Just to align with the previous test (delete and re-add), here we do the
same for "delete re-add signal".

It doesn't seem to be an issue on the Netdev CI for the moment, but
better be safe than sorry. Plus, it doesn't change anything for the
test: the connection will run for longer, but it will be killed when all
operations have been done.

Link: https://lore.kernel.org/r/20240827-selftests-mptcp-longer-tests-v1-2-ec11c05fe980@kernel.org
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>