gh: tests: one LCOV report per branch
It sounds better to unify the reports to see the global code coverage
instead of 3 different reports.

For that, we need to run 'genhtml' after all builds (in case of
success), with MPTCP source code.

While at it, split the handling of flakes and lcov results, and add a
summary for the global coverage.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>