mptcp: change is_backup interfaces as get_flags
The is_backup() interface of path manager is not very common. A more
common approach is to add a get_flags() interface to obtain the flags
value of a given address. Then is_backup() can be implemented through
get_flags() by test whether backup flag is set in the flags value.

Signed-off-by: Geliang Tang <tanggeliang@kylinos.cn>
Message-Id: <33f7ebb53d3195278d42c7b43f5d286e9ea40b4c.1728538976.git.tanggeliang@kylinos.cn>