mptcp: add struct mptcp_id_bitmap
A new struct mptcp_id_bitmap is defined to unify all bitmap type of address
IDs for both in-kernel PM and userspace PM.

This type can be used to easily refactor dump_addr() interface of the path
managers to accept an mptcp_id_bitmap type parameter. It also allows this
parameter of dump_addr() can be modified by BPF program when implementing
this interface of a BFP path manager.

Signed-off-by: Geliang Tang <tanggeliang@kylinos.cn>
Message-Id: <b6f9a6abdfc87858322299c4d4508b233bec8fa1.1728742591.git.tanggeliang@kylinos.cn>