gh: tests: use GitHub registry to avoid rate limit
We already hit Docker hub's registry limits.

Instead of logging there, which would require to set a logging per fork
repo, we can switch to GitHub registry. That's easy, we already already
publishing the images there.

An alternative would be to add a step using the linked Actions, to have
the system downloading the image, and not doing anything with it. But
that doesn't look nice, probably best to avoid that if possible.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>