mptcp: add struct mptcp_id_bitmap
A new struct mptcp_id_bitmap is defined to unify all bitmap type of address
IDs for both in-kernel PM and userspace PM.

This type can be used to easily refactor dump_addr() interface of the path
managers to accept an mptcp_id_bitmap type parameter. It also allows this
parameter of dump_addr() can be modified by BPF program when implementing
this interface of a BFP path manager.

Signed-off-by: Geliang Tang <tanggeliang@kylinos.cn>
Message-Id: <2f13a49c56f06882c5f245585b202825b7db5201.1729069854.git.tanggeliang@kylinos.cn>