mptcp: change is_backup interfaces as get_flags
The is_backup() interface of path manager is not very common. A more
common approach is to add a get_flags() interface to obtain the flags
value of a given address. Then is_backup() can be implemented through
get_flags() by test whether backup flag is set in the flags value.

Signed-off-by: Geliang Tang <tanggeliang@kylinos.cn>
Message-Id: <d52e2d9777a3a35310d12c32a2cc0b58ad7cc206.1729069854.git.tanggeliang@kylinos.cn>