gh: tests: ccache dir with fixed name
The ccache dir for BTF are by default in a different directory. That's
an issue on the CI side because the first time, it took the ccache dir
from another mode, and kept the other directory without using it.

Let's avoid this complexity that is not needed on the CI and have a
fixed ccache dir name.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>