mptcp: update bpf_mptcp_sock_from_subflow
The input parameter of bpf_mptcp_sock_from_subflow() is an msk subsocket
of type IPPROTO_TCP. This patch extends it to accept an IPPROTO_MPTCP
socket. With this change, the function name bpf_mptcp_sock_from_subflow
is no longer appropriate, and it is renamed to bpf_mptcp_sock_from_sock.

Signed-off-by: Geliang Tang <tanggeliang@kylinos.cn>
Message-Id: <fce976c5f62f991f7f6d51a4d08a5f572aec656c.1733724383.git.tanggeliang@kylinos.cn>