mptcp: update bpf_mptcp_sock_from_subflow
The input parameter of bpf_mptcp_sock_from_subflow() is an msk subsocket
of type IPPROTO_TCP. This patch extends it to accept an IPPROTO_MPTCP
socket. With this change, the function name bpf_mptcp_sock_from_subflow
is no longer appropriate, and it is renamed to bpf_mptcp_sock_from_sock.

Signed-off-by: Geliang Tang <tanggeliang@kylinos.cn>
Message-Id: <8f9c34611d46f96a0047928ac41b385faf743805.1733733573.git.tanggeliang@kylinos.cn>