netlink: specs: mptcp: clearly mention attributes
The rendered version of the MPTCP events [1] looked strange, because the
whole content of the 'doc' was displayed in the same block.

It was then not clear that the first words, not even ended by a period,
were the attributes that are defined when such events are emitted. These
attributes have now been moved to the end, prefixed by 'Attributes:' and
ended with a period. Note that '>-' has been added after 'doc:' to allow
':' in the text below.

Fixes: bc8aeb2045e2 ("Documentation: netlink: add a YAML spec for mptcp")
Link: https://docs.kernel.org/networking/netlink_spec/mptcp_pm.html#event-type [1]
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>
Message-Id: <20241216-mptcp-pm-nl-spec-doc-attr-v2-2-68807f55d42c@kernel.org>