mptcp: drop struct mptcp_sched_data
The mptcp_subflow bpf_iter is added now, it's better to use the helper
bpf_for_each(mptcp_subflow) to traverse all subflows on the conn_list of
an MPTCP socket and then call kfunc to modify the fields of each subflow
in the WIP MPTCP BPF packet scheduler examples, instead of converting them
to a fixed array. With this helper, we can get rid of this subflow array
"contexts" and the size of it "subflows" in struct mptcp_sched_data. Then
the entire struct mptcp_sched_data can be dropped.

Signed-off-by: Geliang Tang <tanggeliang@kylinos.cn>
Message-Id: <96d9e097dcdea76f2a86946287de452f0694035b.1734947922.git.tanggeliang@kylinos.cn>