mptcp: change is_backup interfaces as get_flags
is_backup() interface of the path managers is not very common. A more
common approach is to add a get_flags() interface to obtain the flags
value of a given address. Then is_backup() can be implemented through
get_flags() by test whether backup flag is set in the flags value.

Signed-off-by: Geliang Tang <tanggeliang@kylinos.cn>
Message-Id: <95a0cce72e48504cffcb5050a90f2a93d4be5651.1736924314.git.tanggeliang@kylinos.cn>