mptcp: change is_backup interfaces as get_flags
is_backup() interface of the path managers is not very common. A more
common approach is to add a get_flags() interface to obtain the flags
value of a given address. Then is_backup() can be implemented through
get_flags() by test whether backup flag is set in the flags value.

Signed-off-by: Geliang Tang <tanggeliang@kylinos.cn>
Message-Id: <75f5ab33f0a11d32eeaa9af698df6e64bb1da951.1737012165.git.tanggeliang@kylinos.cn>