selftests/bpf: Use pm_nl_ctl if ip mptcp not supported
If running MPTCP BPF selftests on systems with an old version of IPRoute2,
'ip mptcp' command is not supported. In this case, instead of skipping the
test, falling back to using 'pm_nl_ctl' tool is a better option. This patch
adds an 'ip_mptcp' argument for endpoint_add() to control whether to use
'ip mptcp' or 'pm_nl_ctl' to add an endpoint.

Signed-off-by: Geliang Tang <tanggeliang@kylinos.cn>
Message-Id: <6682f174af81e85908db1aff4f379a61d9a77642.1738924354.git.tanggeliang@kylinos.cn>