mptcp: drop subflow contexts in mptcp_sched_data
The mptcp_subflow bpf_iter is added now, it's better to use the helper
bpf_for_each(mptcp_subflow) to traverse all subflows on the conn_list of
an MPTCP socket and then call kfunc to modify the fields of each subflow
in the WIP MPTCP BPF packet scheduler examples, instead of converting them
to a fixed array. With this helper, we can get rid of this subflow array
"contexts" and the size of it "subflows" in struct mptcp_sched_data. And
keep struct mptcp_sched_data for future use.

Signed-off-by: Geliang Tang <tanggeliang@kylinos.cn>
Message-Id: <de2c471ad3f461c54b4da89a72b0b575fbbd4fb7.1739083794.git.tanggeliang@kylinos.cn>