mptcp: sysctl: use index for sysctl table
The sysctl table needs to be updated when a new sysctl is added or an old
sysctl is deleted. It's inconvenient to use fixed index numbers, since the
index number that needs to be changed and all subsequent index numbers need
to be updated accordingly.

This patch adds an "i" variable to access each sysctl in the sysctl table
incrementally.

Signed-off-by: Geliang Tang <tanggeliang@kylinos.cn>
Message-Id: <3e40ffced396775b7e3056b6f0934c112d8e8513.1740320007.git.tanggeliang@kylinos.cn>