mptcp: drop subflow contexts in mptcp_sched_data
The mptcp_subflow_sched bpf_iter is added now, it's better to use the
helper bpf_for_each(mptcp_subflow_sched) to traverse all subflows on the
conn_list of an MPTCP socket and then call kfunc to modify the fields of
each subflow in the WIP MPTCP BPF packet scheduler examples, instead of
converting them to a fixed array. With this helper, we can get rid of
this subflow array "contexts" and the size of it "subflows" in struct
mptcp_sched_data.

Signed-off-by: Geliang Tang <tanggeliang@kylinos.cn>
Message-Id: <54a6a158b655f9e19588a1460867316f6d0cef64.1741347233.git.tanggeliang@kylinos.cn>