mptcp: pm: update pm lock order in mptcp_pm_worker
Later functions that cannot hold the mptcp pm lock will be called from
the PM worker, so this patch modifies the order of holding the lock
at the beginning of this function and releasing the lock at the end.

The new order is to obtain a copy of pm->status while holding the mptcp
pm lock, then read the copy after releasing the lock.

For each PM status flag, hold the lock, clear this flag of pm->status,
and then call each handling function before or after releasing the lock
as needed.

Finally, hold the lock before calling __mptcp_pm_kernel_worker() and
release it afterwards.

Signed-off-by: Geliang Tang <tanggeliang@kylinos.cn>
Message-Id: <739e3667698915fa690d08c751c1846bd9442df2.1743054942.git.tanggeliang@kylinos.cn>