mptcp: pm: add accept_new_address() interface
Similar to the accept_new_subflow() interface, this patch adds a mandatory
accept_new_address() interface for struct mptcp_pm_ops, which is used to
determine whether an ADD_ADDR address is accepted by PM.

This interface is called in mptcp_pm_add_addr_received(). If the address is
not accepted, increase MPTCP_MIB_ADDADDRDROP counter and return.

Signed-off-by: Geliang Tang <tanggeliang@kylinos.cn>
Message-Id: <94e82037a2f86ba85bc0f7c5ada46424b785073b.1743481566.git.tanggeliang@kylinos.cn>