selftests: mptcp: print trailing bytes with od
This is better than printing random bytes in the terminal.

Note Jakub suggested 'hexdump', but Mat found out that this tool is not
often installed by default. 'od' can do a similar job, and it is in the
POSIX specs and available in coreutils, so it should be on more systems.

Suggested-by: Jakub Kicinski <kuba@kernel.org>
Suggested-by: Mat Martineau <martineau@kernel.org>
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>
Message-ID: <20250905-sft-mptcp-disc-err-v2-4-dfb3b6b4a877@kernel.org>