Squash to "mptcp: pm: in-kernel: usable client side with C-flag"
The id_avail_bitmap is only used when either the 'subflow' or 'signal'
flag is used, but not with 'fullmesh' only. Here, it is replacing the
'subflow' action, so check if this flag is set.

Also, add the check for the max subflows upfront, so we avoid scheduling
the worker with MPTCP_PM_ADD_ADDR_RECEIVED if it is not needed.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>
Message-ID: <20250919-pm-kern-endp-add_addr-v2-1-1f8f04f01914@kernel.org>