mptcp: pm: align endpoint flags size with the NL specs
The MPTCP Netlink specs describe the 'flags' as a u32 type. Internally,
a u8 type was used.

Using a u8 is currently fine, because only the 5 first bits are used.
But there is also no reason not to be aligns with the specs, and
to stick to a u8. Especially because there is a whole of 3 bytes after
in both mptcp_pm_local and mptcp_pm_addr_entry structures.

Also, setting it to a u32 will allow future flags, just in case.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>
Message-ID: <20251126-mptcp-pm-kern-drop-unknown-flags-v1-3-d2e4c2ebce0c@kernel.org>