mptcp: cleanup duplicate ack suppression on fastclose.
The current transition to TCP_CLOSE in __mptcp_close_ssk() is rather
obscure, especially since it's done unconditionally for all subflow.

Such action is needed only when disposing the given socket - that is
for MPJ subflows, or at mptcp close time.

Move the statement in a better place and explain it's role for future
memory.

Signed-off-by: Paolo Abeni <pabeni@redhat.com>
Message-ID: <6d4733b0268246892cbbb8c2d3bcebce1714404a.1764786299.git.pabeni@redhat.com>