gh: new job to validate CI kernels
Some CIs like LKFT (thanks for that!) are regularly validating the
stable kernels, and they include our MPTCP selftests.

Unfortunately, new issues are not always reported, tests are executed in
VMs without KVM support (slow, some tests don't pass because of that),
and MPTCP packetdrill tests are not executed.

With this new job, the validation and tracking is now done on our side.
By default, all supported kernels are tests in parallel. Queued patches
-- from 'stable/stable-queue' repo -- and pending ones -- from MPTCP
patchwork, "Handled Elsewhere" status with a version prefix -- are
applied . Regarding the tests modes, only the 'normal' and 'debug' ones
are used, assuming BPF tests are handled by BPF maintainers.

Note that no cache is used, because the cache size is limited on the
public CI. That's fine, we have time for these tests.

Results are published on IRC, and on a new folder on the flakes side.

Closes: https://github.com/multipath-tcp/mptcp_net-next/issues/604
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>