gh: stable: run latest kselftests once a week
According to the doc [1], kselftests are supposed to run on any kernels,
including the old ones not supporting new features. Some CI like LKFT
are always using the kselftests from the last stable version.

It is then important to check that new tests (from the export branch)
validating new features are correctly skipped on older kernels.

This is what is being done here: kselftests from the export branches are
downloaded, built and moved to replace the ones from the stable version.

Results are sent in a different directory.

Link: https://docs.kernel.org/dev-tools/kselftest.html [1]
Closes: https://github.com/multipath-tcp/mptcp_net-next/issues/396
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>